+++ R/SingleGroup-methods.R
@@ -461,7 +461,7 @@ setMethod(
 #' residuals(x, suppress = .15)
 #'
 #' # with and without supplied factor scores
-#' Theta <- fscores(x, full.scores=TRUE)
+#' Theta <- fscores(x)
 #' residuals(x, type = 'Q3', Theta=Theta)
 #' residuals(x, type = 'Q3', method = 'ML')
 #'