+++ R/fscores.R
@@ -10,7 +10,7 @@
 #' information matrix was computed, which are useful if the sample size/number of items were small.
 #' As well, if the model contained latent regression predictors this information will
 #' be used in computing MAP and EAP estimates (for these models, \code{full.scores=TRUE}
-#' by default). Finally, plausible value imputation is also available, and will also account
+#' will always be used). Finally, plausible value imputation is also available, and will also account
 #' for latent regression predictor effects.
 #'
 #' The function will return either a table with the computed scores and standard errors,
@@ -29,9 +29,9 @@
 #' @aliases fscores
 #' @param object a computed model object of class \code{SingleGroupClass},
 #'   \code{MultipleGroupClass}, or \code{DiscreteClass}
-#' @param full.scores if \code{FALSE} (default) then a summary table with
-#'   factor scores for each unique pattern is displayed. Otherwise a matrix of factor scores
-#'   for each response pattern in the data is returned
+#' @param full.scores if \code{FALSE} then a summary table with
+#'   factor scores for each unique pattern is displayed. Otherwise, a matrix of factor scores
+#'   for each response pattern in the data is returned (default)
 #' @param rotate prior rotation to be used when estimating the factor scores. See
 #'   \code{\link{summary-method}} for details. If the object is not an exploratory model
 #'   then this argument is ignored
@@ -110,15 +110,15 @@
 #' \dontrun{
 #'
 #' mod <- mirt(Science, 1)
-#' tabscores <- fscores(mod)
+#' tabscores <- fscores(mod, full.scores = FALSE)
 #' head(tabscores)
-#' fullscores <- fscores(mod, full.scores = TRUE)
-#' fullscores_with_SE <- fscores(mod, full.scores = TRUE, full.scores.SE=TRUE)
+#' fullscores <- fscores(mod)
+#' fullscores_with_SE <- fscores(mod, full.scores.SE=TRUE)
 #' head(fullscores)
 #' head(fullscores_with_SE)
 #'
 #' #change method argument to use MAP estimates
-#' fullscores <- fscores(mod, full.scores = TRUE, method='MAP')
+#' fullscores <- fscores(mod, method='MAP')
 #' head(fullscores)
 #'
 #' #calculate MAP for a given response vector
@@ -127,12 +127,15 @@
 #' fscores(mod, method='MAP', response.pattern = rbind(c(1,2,3,4), c(2,2,1,3)))
 #'
 #' #use custom latent variable properties (diffuse prior for MAP is very close to ML)
-#' fscores(mod, method='MAP', cov = matrix(1000))
-#' fscores(mod, method='ML')
+#' fscores(mod, method='MAP', cov = matrix(1000), full.scores = FALSE)
+#' fscores(mod, method='ML', full.scores = FALSE)
+#'
+#' # EAPsum table of values based on total scores
+#' fscores(mod, method = 'EAPsum', full.scores = FALSE)
 #'
 #' #WLE estimation, run in parallel using available cores
 #' mirtCluster()
-#' fscores(mod, method='WLE')
+#' fscores(mod, method='WLE', full.scores = FALSE)
 #'
 #' #multiple imputation using 30 draws for EAP scores. Requires information matrix
 #' mod <- mirt(Science, 1, SE=TRUE)
@@ -150,10 +153,10 @@
 #' library(msm)
 #' # need the :: scope for parallel to see the function (not require if no mirtCluster() defined)
 #' fun <- function(Theta, ...) msm::dtnorm(Theta, mean = 0, sd = 1, lower = -2, upper = 5)
-#' head(fscores(mod, custom_den = fun, method = 'MAP'))
+#' head(fscores(mod, custom_den = fun, method = 'MAP', full.scores = FALSE))
 #'
 #'}
-fscores <- function(object, rotate = 'oblimin', Target = NULL, full.scores = FALSE, method = "EAP",
+fscores <- function(object, rotate = 'oblimin', Target = NULL, full.scores = TRUE, method = "EAP",
                     quadpts = NULL, response.pattern = NULL, plausible.draws = 0,
                     returnER = FALSE, return.acov = FALSE, mean = NULL, cov = NULL, verbose = TRUE,
                     full.scores.SE = FALSE, theta_lim = c(-6,6), MI = 0,
@@ -170,6 +173,7 @@ fscores <- function(object, rotate = 'oblimin', Target = NULL, full.scores = FAL
         plausible.draws <- 1
         method <- 'EAP'
     }
+    if(returnER) full.scores <- FALSE
     ret <- fscores.internal(object=object, rotate=rotate, full.scores=full.scores, method=method,
                             quadpts=quadpts, response.pattern=response.pattern, QMC=QMC,
                             verbose=verbose, returnER=returnER, gmean=mean, gcov=cov,