+++ R/itemfit.R
@@ -94,7 +94,7 @@
 #' itemfit(raschfit, method = 'ML') #infit and outfit stats
 #'
 #' #same as above, but inputting ML estimates instead
-#' Theta <- fscores(raschfit, method = 'ML', full.scores=TRUE)
+#' Theta <- fscores(raschfit, method = 'ML')
 #' itemfit(raschfit, Theta=Theta)
 #'
 #' #similar example to Kang and Chen 2007