+++ tests/extratests/test-02-wideData.R
@@ -15,7 +15,7 @@ test_that('wide1dim', {
     vals <- vals[vals != 0 & vals != 1]
     expect_equal(fivenum(vals), c(-3.50602785, 0.03518035, 0.90189515, 1.01846863, 3.46084417),
                  tolerance = 1e-4)
-    EAP <- fscores(mod, verbose=FALSE)
+    EAP <- fscores(mod, verbose=FALSE, full.scores=FALSE)
     expect_equal(head(EAP[,'F1']), c(-1.4040, -2.5166, -1.5617, -1.8788, -2.5433, -1.1932),
                  tolerance=1e-4)
 })